add test tags for unit and integration tests
