+++ tests/pytest.ini
@@ -1,0 +1,8 @@
+[pytest]
+markers =
+    unit: marks tests as unit (deselect with '-m "not unit"')
+    serial
+    integration: marks tests as integration (deselect with '-m "not integration"')
+    serial
+    databricks: marks tests as requires databricks (deselect with '-m "not databricks"')
+    serial
